separate validate datasets loop
