+++ bin/check-input-datasets.js
@@ -1,20 +1,9 @@
 const fsPromises = require("fs").promises;
-const path = require("path");
-const {
-  getInputDatasetSchema,
-} = require("../src/data-validation/get-input-dataset-schema");
 const { DATA_FOLDER_NAME } = require("../src/process-single-dataset/constants");
-const utils = require("../src/utils");
-const dataPrep = require("../src/data-validation/data-prep");
-const unpackInputDataset = require("../src/data-validation/unpack-input-dataset");
-const {checkForError, checkSingleDatasetInput, datasetCheck} = require("./check-single-dataset");
-// referenced partial schemas
-const INPUT_DATASET_SCHEMA_FILE = "input-dataset.schema.json";
-const INPUT_MEGASET_SCHEMA_FILE = "input-megaset.schema.json";
-const ajv = getInputDatasetSchema();
+const {validateSingleDataset} = require("./check-single-dataset");
 
 
-const validateDatasets = (singleDatasetFolder = null) => {
+const validateDatasets = () => {
   fsPromises
     .readdir(`./${DATA_FOLDER_NAME}`)
     .then(async (files) => {
@@ -33,42 +22,7 @@ const validateDatasets = (singleDatasetFolder = null) => {
         }
       }
       for (const datasetFolder of foldersToCheck) {
-        // skip this dataset if singleDatasetFolder is specified
-        if (singleDatasetFolder && path.basename(datasetFolder) !== path.basename(singleDatasetFolder)) { 
-          continue; 
-        } else if (singleDatasetFolder) {
-          console.log(`logging featureName and order... ${datasetFolder}`);
-          datasetCheck();
-        }; 
-        const topLevelJson = await utils.readDatasetJson(datasetFolder);
-        if (topLevelJson.datasets) {
-          // is a megaset, need to check both megaset
-          // file and each dataset folder
-          const foundError = checkForError(
-            `${datasetFolder}/dataset.json`,
-            topLevelJson,
-            INPUT_MEGASET_SCHEMA_FILE
-          );
-          if (foundError) {
-            hasError = true;
-          }
-          for (const subDatasetFolder of topLevelJson.datasets) {
-            const datasetReadFolder = `${datasetFolder}/${subDatasetFolder}`;
-            if (singleDatasetFolder) {
-              console.log(`logging featureName and order... ${datasetFolder}`);
-              datasetCheck();
-            };
-            const foundSubError = await checkSingleDatasetInput(datasetReadFolder);
-            if (foundSubError) {
-              hasError = true;
-            }
-          }
-        } else {
-          const foundError = await checkSingleDatasetInput(datasetFolder);
-          if (foundError) {
-            hasError = true;
-          }
-        }
+        hasError = await validateSingleDataset(datasetFolder);
       }
       return hasError;
     })
@@ -80,8 +34,5 @@ const validateDatasets = (singleDatasetFolder = null) => {
     });
 };
 
-if (process.argv[2]) {
-  validateDatasets(process.argv[2]);
-} else {
+
   validateDatasets();
-}
